refactor(relayer): move getMessageHash to utils file
