+++ src/relayer/relayer.service.ts
@@ -7,15 +7,14 @@ import {
   RequestWithdrawalAtBlocks,
 } from './relayer.interface';
 import { MulticallRequest, MulticallResponse } from 'web3/web3.interface';
-import { TRANSFER_FROM_STARKNET, ZeroBytes, l2BridgeAddressToL1 } from './relayer.constants';
+import { ZeroBytes, l2BridgeAddressToL1 } from './relayer.constants';
 import { MongoService } from 'storage/mongo/mongo.service';
-import { ethers } from 'ethers';
-import { uint256 } from 'starknet';
 import { Transfer, Withdrawal } from 'indexer/entities';
 import { IndexerService } from 'indexer/indexer.service';
 import { callWithRetry, sleep } from './relayer.utils';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { PrometheusService } from 'common/prometheus';
+import { getMessageHash } from './utils';
 
 @Injectable()
 export class RelayerService {
@@ -196,7 +195,7 @@ export class RelayerService {
         multicallRequests.push({
           target: l2BridgeAddressToL1Addresses[withdrawal.bridgeAddress].l1BridgeAddress,
           callData: this.web3Service.encodeCalldataStarknetCore('l2ToL1Messages', [
-            this.getMessageHash(withdrawal.bridgeAddress, l1BridgeAddress, withdrawal.l1Recipient, withdrawal.amount),
+            getMessageHash(withdrawal.bridgeAddress, l1BridgeAddress, withdrawal.l1Recipient, withdrawal.amount),
           ]),
         });
       }
@@ -273,15 +272,6 @@ export class RelayerService {
         throw errMessage;
       },
     });
-  }
-
-  getMessageHash(l2BridgeAddress: string, l1BridgeAddress: string, receiverL1: string, amount: string): string {
-    const amountUint256 = uint256.bnToUint256(amount.toString());
-    const payload = [TRANSFER_FROM_STARKNET, receiverL1, amountUint256.low, amountUint256.high];
-    return ethers.utils.solidityKeccak256(
-      ['uint256', 'uint256', 'uint256', 'uint256[]'],
-      [l2BridgeAddress, l1BridgeAddress, payload.length, payload],
-    );
   }
 
   async canProcessWithdrawals(): Promise<CheckCanProcessWithdrawalsResults> {
